chore: organização da validação de data de vencimento na rota de dashboard
